LPS-50491 - Move liferayselenium impl to portalweb/util/liferayselenium/impl
