+++ portal-web/test/functional/com/liferay/portalweb/util/liferayselenium/impl/AppiumMobileDriverImpl.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portalweb.portal.util.liferayselenium;
+package com.liferay.portalweb.util.liferayselenium.impl;
 
 import io.appium.java_client.AppiumDriver;
